# docs/openapi.yml
openapi: 3.0.3
info:
  title: Team Dev Server API
  description: |-
  version: 1.0
servers:
  - url: http://localhost:4000/
...
paths:
  /loggedInUser:
    put:
      tags:
        - User
      summary: Update logged-in user
      description: Updates the information of the logged-in user.
      operationId: updateLoggedInUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                bio:
                  type: string
                githubUrl:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '201':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        cohortId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        githubUrl:
          type: string
...